+++ client/src/components/fileLibrary/constants.js
@@ -1,12 +1,25 @@
-import { THREE_D_MODEL_LABEL } from '../../utils/cadFileTypes';
+import {
+  getPspiceFileLabel,
+  PSPICE_LABEL,
+  SCHEMATIC_SYMBOL_LABEL,
+  THREE_D_MODEL_LABEL,
+} from '../../utils/cadFileTypes';
 
 // Map file type IDs from cad_files.file_type to display labels
 export const fileTypeLabels = {
   footprint: 'PCB Footprint',
-  symbol: 'Schematic',
+  symbol: SCHEMATIC_SYMBOL_LABEL,
   model: THREE_D_MODEL_LABEL,
-  pspice: 'PSpice Model',
+  pspice: PSPICE_LABEL,
   pad: 'Pad File',
+};
+
+export const getCadFileTypeLabel = (fileType, fileName) => {
+  if (fileType === 'pspice') {
+    return getPspiceFileLabel(fileName);
+  }
+
+  return fileTypeLabels[fileType] || fileType;
 };
 
 // Map route type IDs to cad_files.file_type values