+++ server/src/constants/cadFiles.js
@@ -8,3 +8,20 @@ export const MODEL_FILE_EXTENSIONS = Object.freeze([
   '.3ds',
   '.x_t',
 ]);
+
+export const PSPICE_MODEL_FILE_EXTENSIONS = Object.freeze([
+  '.lib',
+  '.cir',
+  '.sub',
+  '.inc',
+  '.mod',
+]);
+
+export const PSPICE_SYMBOL_FILE_EXTENSIONS = Object.freeze([
+  '.olb',
+]);
+
+export const PSPICE_FILE_EXTENSIONS = Object.freeze([
+  ...PSPICE_MODEL_FILE_EXTENSIONS,
+  ...PSPICE_SYMBOL_FILE_EXTENSIONS,
+]);